Fixes after rebasing

+++ interfaces/src/main/java/ar/edu/itba/paw/persistance/ProductDao.java
@@ -17,6 +17,6 @@ public interface ProductDao {
 
     boolean delete(long productId);
 
-    double getRestaurantAveragePrice(int restaurantId);
+    double getRestaurantAveragePrice(long restaurantId);
 
 }
