+++ persistence/src/main/java/ar/edu/itba/paw/persistence/ProductJdbcDao.java
@@ -90,7 +90,7 @@ public class ProductJdbcDao implements ProductDao {
     }
 
     @Override
-    public double getRestaurantAveragePrice(int restaurantId) {
+    public double getRestaurantAveragePrice(long restaurantId) {
         return jdbcTemplate.queryForObject(
                 "SELECT COALESCE(AVG(price), 0) FROM products JOIN categories ON products.category_id = categories.category_id WHERE categories.restaurant_id = ? AND products.available = true",
                 Double.class,
